LayoutTests/fast/selectors/case-insensitive-attribute-register-requirement-2-expected.txt

Check various cases of register pressure with case-insensitive attribute matching (e.g. [foo=bar i]).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing "[type=foobar i] > * > * >> innerTestCase"
PASS document.querySelectorAll('[type=foobar i] > * > * >> innerTestCase').length is 8
PASS document.querySelectorAll('[type=foobar i] > * > * >> innerTestCase')[0].id is "innertestcase1"
PASS document.querySelectorAll('[type=foobar i] > * > * >> innerTestCase')[1].id is "innertestcase2"
PASS document.querySelectorAll('[type=foobar i] > * > * >> innerTestCase')[2].id is "innertestcase3"
PASS document.querySelectorAll('[type=foobar i] > * > * >> innerTestCase')[3].id is "innertestcase4"
PASS document.querySelectorAll('[type=foobar i] > * > * >> innerTestCase')[4].id is "innertestcase5"
PASS document.querySelectorAll('[type=foobar i] > * > * >> innerTestCase')[5].id is "innertestcase6"
PASS document.querySelectorAll('[type=foobar i] > * > * >> innerTestCase')[6].id is "innertestcase7"
PASS document.querySelectorAll('[type=foobar i] > * > * >> innerTestCase')[7].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[attribute1=VALUE1 i] > * > * >> innerTestCase"
PASS document.querySelectorAll('[attribute1=VALUE1 i] > * > * >> innerTestCase').length is 7
PASS document.querySelectorAll('[attribute1=VALUE1 i] > * > * >> innerTestCase')[0].id is "innertestcase1"
PASS document.querySelectorAll('[attribute1=VALUE1 i] > * > * >> innerTestCase')[1].id is "innertestcase2"
PASS document.querySelectorAll('[attribute1=VALUE1 i] > * > * >> innerTestCase')[2].id is "innertestcase3"
PASS document.querySelectorAll('[attribute1=VALUE1 i] > * > * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[attribute1=VALUE1 i] > * > * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[attribute1=VALUE1 i] > * > * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[attribute1=VALUE1 i] > * > * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type=foobar i] ~ * ~ * >> innerTestCase"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type=foobar i] ~ * ~ * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[attribute1=VALUE1 i] ~ * ~ * >> innerTestCase"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ * >> innerTestCase').length is 7
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type=foobar i] ~ * ~ *  > * > * >> innerTestCase"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ *  > * > * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type=foobar i] ~ * ~ *  > * > * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ *  > * > * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ *  > * > * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ *  > * > * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ *  > * > * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ *  > * > * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type=foobar i] ~ * ~ *  > * > * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[attribute1=VALUE1 i] ~ * ~ *  > * > * >> innerTestCase"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ *  > * > * >> innerTestCase').length is 7
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ *  > * > * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ *  > * > * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ *  > * > * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ *  > * > * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ *  > * > * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ *  > * > * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[attribute1=VALUE1 i] ~ * ~ *  > * > * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type~=fOobar i] > * > * >> innerTestCase"
PASS document.querySelectorAll('[type~=fOobar i] > * > * >> innerTestCase').length is 8
PASS document.querySelectorAll('[type~=fOobar i] > * > * >> innerTestCase')[0].id is "innertestcase1"
PASS document.querySelectorAll('[type~=fOobar i] > * > * >> innerTestCase')[1].id is "innertestcase2"
PASS document.querySelectorAll('[type~=fOobar i] > * > * >> innerTestCase')[2].id is "innertestcase3"
PASS document.querySelectorAll('[type~=fOobar i] > * > * >> innerTestCase')[3].id is "innertestcase4"
PASS document.querySelectorAll('[type~=fOobar i] > * > * >> innerTestCase')[4].id is "innertestcase5"
PASS document.querySelectorAll('[type~=fOobar i] > * > * >> innerTestCase')[5].id is "innertestcase6"
PASS document.querySelectorAll('[type~=fOobar i] > * > * >> innerTestCase')[6].id is "innertestcase7"
PASS document.querySelectorAll('[type~=fOobar i] > * > * >> innerTestCase')[7].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type~=foObar i] ~ * ~ * >> innerTestCase"
PASS document.querySelectorAll('[type~=foObar i] ~ * ~ * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type~=foObar i] ~ * ~ * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type~=foObar i] ~ * ~ * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type~=foObar i] ~ * ~ * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type~=foObar i] ~ * ~ * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type~=foObar i] ~ * ~ * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type~=foObar i] ~ * ~ * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type~=foObar i] ~ * ~ * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type~=foobar i] ~ * ~ *  > * > * >> innerTestCase"
PASS document.querySelectorAll('[type~=foobar i] ~ * ~ *  > * > * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type~=foobar i] ~ * ~ *  > * > * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type~=foobar i] ~ * ~ *  > * > * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type~=foobar i] ~ * ~ *  > * > * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type~=foobar i] ~ * ~ *  > * > * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type~=foobar i] ~ * ~ *  > * > * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type~=foobar i] ~ * ~ *  > * > * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type~=foobar i] ~ * ~ *  > * > * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type|=foObar i] > * > * >> innerTestCase"
PASS document.querySelectorAll('[type|=foObar i] > * > * >> innerTestCase').length is 8
PASS document.querySelectorAll('[type|=foObar i] > * > * >> innerTestCase')[0].id is "innertestcase1"
PASS document.querySelectorAll('[type|=foObar i] > * > * >> innerTestCase')[1].id is "innertestcase2"
PASS document.querySelectorAll('[type|=foObar i] > * > * >> innerTestCase')[2].id is "innertestcase3"
PASS document.querySelectorAll('[type|=foObar i] > * > * >> innerTestCase')[3].id is "innertestcase4"
PASS document.querySelectorAll('[type|=foObar i] > * > * >> innerTestCase')[4].id is "innertestcase5"
PASS document.querySelectorAll('[type|=foObar i] > * > * >> innerTestCase')[5].id is "innertestcase6"
PASS document.querySelectorAll('[type|=foObar i] > * > * >> innerTestCase')[6].id is "innertestcase7"
PASS document.querySelectorAll('[type|=foObar i] > * > * >> innerTestCase')[7].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type|=fooBar i] ~ * ~ * >> innerTestCase"
PASS document.querySelectorAll('[type|=fooBar i] ~ * ~ * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type|=fooBar i] ~ * ~ * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type|=fooBar i] ~ * ~ * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type|=fooBar i] ~ * ~ * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type|=fooBar i] ~ * ~ * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type|=fooBar i] ~ * ~ * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type|=fooBar i] ~ * ~ * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type|=fooBar i] ~ * ~ * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type|=foobar i] ~ * ~ *  > * > * >> innerTestCase"
PASS document.querySelectorAll('[type|=foobar i] ~ * ~ *  > * > * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type|=foobar i] ~ * ~ *  > * > * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type|=foobar i] ~ * ~ *  > * > * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type|=foobar i] ~ * ~ *  > * > * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type|=foobar i] ~ * ~ *  > * > * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type|=foobar i] ~ * ~ *  > * > * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type|=foobar i] ~ * ~ *  > * > * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type|=foobar i] ~ * ~ *  > * > * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type^=fooBa i] > * > * >> innerTestCase"
PASS document.querySelectorAll('[type^=fooBa i] > * > * >> innerTestCase').length is 8
PASS document.querySelectorAll('[type^=fooBa i] > * > * >> innerTestCase')[0].id is "innertestcase1"
PASS document.querySelectorAll('[type^=fooBa i] > * > * >> innerTestCase')[1].id is "innertestcase2"
PASS document.querySelectorAll('[type^=fooBa i] > * > * >> innerTestCase')[2].id is "innertestcase3"
PASS document.querySelectorAll('[type^=fooBa i] > * > * >> innerTestCase')[3].id is "innertestcase4"
PASS document.querySelectorAll('[type^=fooBa i] > * > * >> innerTestCase')[4].id is "innertestcase5"
PASS document.querySelectorAll('[type^=fooBa i] > * > * >> innerTestCase')[5].id is "innertestcase6"
PASS document.querySelectorAll('[type^=fooBa i] > * > * >> innerTestCase')[6].id is "innertestcase7"
PASS document.querySelectorAll('[type^=fooBa i] > * > * >> innerTestCase')[7].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type^=fOoba i] ~ * ~ * >> innerTestCase"
PASS document.querySelectorAll('[type^=fOoba i] ~ * ~ * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type^=fOoba i] ~ * ~ * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type^=fOoba i] ~ * ~ * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type^=fOoba i] ~ * ~ * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type^=fOoba i] ~ * ~ * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type^=fOoba i] ~ * ~ * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type^=fOoba i] ~ * ~ * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type^=fOoba i] ~ * ~ * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type^=fooba i] ~ * ~ *  > * > * >> innerTestCase"
PASS document.querySelectorAll('[type^=fooba i] ~ * ~ *  > * > * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type^=fooba i] ~ * ~ *  > * > * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type^=fooba i] ~ * ~ *  > * > * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type^=fooba i] ~ * ~ *  > * > * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type^=fooba i] ~ * ~ *  > * > * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type^=fooba i] ~ * ~ *  > * > * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type^=fooba i] ~ * ~ *  > * > * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type^=fooba i] ~ * ~ *  > * > * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type$=oobAr i] > * > * >> innerTestCase"
PASS document.querySelectorAll('[type$=oobAr i] > * > * >> innerTestCase').length is 8
PASS document.querySelectorAll('[type$=oobAr i] > * > * >> innerTestCase')[0].id is "innertestcase1"
PASS document.querySelectorAll('[type$=oobAr i] > * > * >> innerTestCase')[1].id is "innertestcase2"
PASS document.querySelectorAll('[type$=oobAr i] > * > * >> innerTestCase')[2].id is "innertestcase3"
PASS document.querySelectorAll('[type$=oobAr i] > * > * >> innerTestCase')[3].id is "innertestcase4"
PASS document.querySelectorAll('[type$=oobAr i] > * > * >> innerTestCase')[4].id is "innertestcase5"
PASS document.querySelectorAll('[type$=oobAr i] > * > * >> innerTestCase')[5].id is "innertestcase6"
PASS document.querySelectorAll('[type$=oobAr i] > * > * >> innerTestCase')[6].id is "innertestcase7"
PASS document.querySelectorAll('[type$=oobAr i] > * > * >> innerTestCase')[7].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type$=oObar i] ~ * ~ * >> innerTestCase"
PASS document.querySelectorAll('[type$=oObar i] ~ * ~ * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type$=oObar i] ~ * ~ * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type$=oObar i] ~ * ~ * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type$=oObar i] ~ * ~ * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type$=oObar i] ~ * ~ * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type$=oObar i] ~ * ~ * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type$=oObar i] ~ * ~ * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type$=oObar i] ~ * ~ * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type$=oobar i] ~ * ~ *  > * > * >> innerTestCase"
PASS document.querySelectorAll('[type$=oobar i] ~ * ~ *  > * > * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type$=oobar i] ~ * ~ *  > * > * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type$=oobar i] ~ * ~ *  > * > * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type$=oobar i] ~ * ~ *  > * > * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type$=oobar i] ~ * ~ *  > * > * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type$=oobar i] ~ * ~ *  > * > * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type$=oobar i] ~ * ~ *  > * > * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type$=oobar i] ~ * ~ *  > * > * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type*=oOba i] > * > * >> innerTestCase"
PASS document.querySelectorAll('[type*=oOba i] > * > * >> innerTestCase').length is 8
PASS document.querySelectorAll('[type*=oOba i] > * > * >> innerTestCase')[0].id is "innertestcase1"
PASS document.querySelectorAll('[type*=oOba i] > * > * >> innerTestCase')[1].id is "innertestcase2"
PASS document.querySelectorAll('[type*=oOba i] > * > * >> innerTestCase')[2].id is "innertestcase3"
PASS document.querySelectorAll('[type*=oOba i] > * > * >> innerTestCase')[3].id is "innertestcase4"
PASS document.querySelectorAll('[type*=oOba i] > * > * >> innerTestCase')[4].id is "innertestcase5"
PASS document.querySelectorAll('[type*=oOba i] > * > * >> innerTestCase')[5].id is "innertestcase6"
PASS document.querySelectorAll('[type*=oOba i] > * > * >> innerTestCase')[6].id is "innertestcase7"
PASS document.querySelectorAll('[type*=oOba i] > * > * >> innerTestCase')[7].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type*=ooBa i] ~ * ~ * >> innerTestCase"
PASS document.querySelectorAll('[type*=ooBa i] ~ * ~ * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type*=ooBa i] ~ * ~ * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type*=ooBa i] ~ * ~ * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type*=ooBa i] ~ * ~ * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type*=ooBa i] ~ * ~ * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type*=ooBa i] ~ * ~ * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type*=ooBa i] ~ * ~ * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type*=ooBa i] ~ * ~ * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

Testing "[type*=ooba i] ~ * ~ *  > * > * >> innerTestCase"
PASS document.querySelectorAll('[type*=ooba i] ~ * ~ *  > * > * >> innerTestCase').length is 7
PASS document.querySelectorAll('[type*=ooba i] ~ * ~ *  > * > * >> innerTestCase')[0].id is "innertestcase2"
PASS document.querySelectorAll('[type*=ooba i] ~ * ~ *  > * > * >> innerTestCase')[1].id is "innertestcase3"
PASS document.querySelectorAll('[type*=ooba i] ~ * ~ *  > * > * >> innerTestCase')[2].id is "innertestcase4"
PASS document.querySelectorAll('[type*=ooba i] ~ * ~ *  > * > * >> innerTestCase')[3].id is "innertestcase5"
PASS document.querySelectorAll('[type*=ooba i] ~ * ~ *  > * > * >> innerTestCase')[4].id is "innertestcase6"
PASS document.querySelectorAll('[type*=ooba i] ~ * ~ *  > * > * >> innerTestCase')[5].id is "innertestcase7"
PASS document.querySelectorAll('[type*=ooba i] ~ * ~ *  > * > * >> innerTestCase')[6].id is "innertestcase8"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[4]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("innerTestCase")[7]).backgroundColor is "rgb(10, 100, 200)"

PASS successfullyParsed is true

TEST COMPLETE
